requirements.txt
flask
requests
gtts
SpeechRecognition
pydub
telebot
python-dotenv
deep-translator